+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Change Log
 
  *  Feature: Add special handling for font assets
  *  Fix: Validate Crate JAR hash and force mergeAssets task to run if changed
+ *  Improvement: Migrate java class generation to JavaPoet
 
 
 ## Version 0.1